bump sql v1.3

+++ NekoRobot/modules/sql/cust_filters_sql.py
@@ -25,7 +25,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import threading
 
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func
+from sqlalchemy import Column, String, UnicodeText, Boolean, BigInteger, distinct, func
 
 from NekoRobot.modules.helper_funcs.msg_types import Types
 from NekoRobot.modules.sql import BASE, SESSION
@@ -49,10 +49,10 @@ class CustomFilters(BASE):
 
     # NEW FILTER
     # alter table cust_filters add column reply_text text;
-    # alter table cust_filters add column file_type integer default 1;
+    # alter table cust_filters add column file_type Biginteger default 1;
     # alter table cust_filters add column file_id text;
     reply_text = Column(UnicodeText)
-    file_type = Column(Integer, nullable=False, default=1)
+    file_type = Column(BigInteger, nullable=False, default=1)
     file_id = Column(UnicodeText, default=None)
 
     def __init__(
@@ -92,10 +92,8 @@ class CustomFilters(BASE):
 
     def __eq__(self, other):
         return bool(
-            isinstance(other, CustomFilters)
-            and self.chat_id == other.chat_id
-            and self.keyword == other.keyword
-        )
+            isinstance(other, CustomFilters) and
+            self.chat_id == other.chat_id and self.keyword == other.keyword)
 
 
 class NewCustomFilters(BASE):
@@ -103,7 +101,7 @@ class NewCustomFilters(BASE):
     chat_id = Column(String(14), primary_key=True)
     keyword = Column(UnicodeText, primary_key=True, nullable=False)
     text = Column(UnicodeText)
-    file_type = Column(Integer, nullable=False, default=1)
+    file_type = Column(BigInteger, nullable=False, default=1)
     file_id = Column(UnicodeText, default=None)
 
     def __init__(self, chat_id, keyword, text, file_type, file_id):
@@ -118,15 +116,13 @@ class NewCustomFilters(BASE):
 
     def __eq__(self, other):
         return bool(
-            isinstance(other, CustomFilters)
-            and self.chat_id == other.chat_id
-            and self.keyword == other.keyword
-        )
+            isinstance(other, CustomFilters) and
+            self.chat_id == other.chat_id and self.keyword == other.keyword)
 
 
 class Buttons(BASE):
     __tablename__ = "cust_filter_urls"
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     chat_id = Column(String(14), primary_key=True)
     keyword = Column(UnicodeText, primary_key=True)
     name = Column(UnicodeText, nullable=False)
@@ -178,10 +174,9 @@ def add_filter(
         if prev:
             with BUTTON_LOCK:
                 prev_buttons = (
-                    SESSION.query(Buttons)
-                    .filter(Buttons.chat_id == str(chat_id), Buttons.keyword == keyword)
-                    .all()
-                )
+                    SESSION.query(Buttons).filter(
+                        Buttons.chat_id == str(chat_id),
+                        Buttons.keyword == keyword).all())
                 for btn in prev_buttons:
                     SESSION.delete(btn)
             SESSION.delete(prev)
@@ -223,10 +218,9 @@ def new_add_filter(chat_id, keyword, reply_text, file_type, file_id, buttons):
         if prev:
             with BUTTON_LOCK:
                 prev_buttons = (
-                    SESSION.query(Buttons)
-                    .filter(Buttons.chat_id == str(chat_id), Buttons.keyword == keyword)
-                    .all()
-                )
+                    SESSION.query(Buttons).filter(
+                        Buttons.chat_id == str(chat_id),
+                        Buttons.keyword == keyword).all())
                 for btn in prev_buttons:
                     SESSION.delete(btn)
             SESSION.delete(prev)
@@ -270,10 +264,9 @@ def remove_filter(chat_id, keyword):
 
             with BUTTON_LOCK:
                 prev_buttons = (
-                    SESSION.query(Buttons)
-                    .filter(Buttons.chat_id == str(chat_id), Buttons.keyword == keyword)
-                    .all()
-                )
+                    SESSION.query(Buttons).filter(
+                        Buttons.chat_id == str(chat_id),
+                        Buttons.keyword == keyword).all())
                 for btn in prev_buttons:
                     SESSION.delete(btn)
 
@@ -291,13 +284,10 @@ def get_chat_triggers(chat_id):
 
 def get_chat_filters(chat_id):
     try:
-        return (
-            SESSION.query(CustomFilters)
-            .filter(CustomFilters.chat_id == str(chat_id))
-            .order_by(func.length(CustomFilters.keyword).desc())
-            .order_by(CustomFilters.keyword.asc())
-            .all()
-        )
+        return (SESSION.query(CustomFilters).filter(
+            CustomFilters.chat_id == str(chat_id)).order_by(
+                func.length(CustomFilters.keyword).desc()).order_by(
+                    CustomFilters.keyword.asc()).all())
     finally:
         SESSION.close()
 
@@ -318,12 +308,9 @@ def add_note_button_to_db(chat_id, keyword, b_name, url, same_line):
 
 def get_buttons(chat_id, keyword):
     try:
-        return (
-            SESSION.query(Buttons)
-            .filter(Buttons.chat_id == str(chat_id), Buttons.keyword == keyword)
-            .order_by(Buttons.id)
-            .all()
-        )
+        return (SESSION.query(Buttons).filter(
+            Buttons.chat_id == str(chat_id),
+            Buttons.keyword == keyword).order_by(Buttons.id).all())
     finally:
         SESSION.close()
 
@@ -337,7 +324,8 @@ def num_filters():
 
 def num_chats():
     try:
-        return SESSION.query(func.count(distinct(CustomFilters.chat_id))).scalar()
+        return SESSION.query(func.count(distinct(
+            CustomFilters.chat_id))).scalar()
     finally:
         SESSION.close()
 
@@ -385,12 +373,10 @@ def __migrate_filters():
             print(str(x.chat_id), x.keyword, x.reply, file_type.value)
             if file_type == Types.TEXT:
                 filt = CustomFilters(
-                    str(x.chat_id), x.keyword, x.reply, file_type.value, None
-                )
+                    str(x.chat_id), x.keyword, x.reply, file_type.value, None)
             else:
                 filt = CustomFilters(
-                    str(x.chat_id), x.keyword, None, file_type.value, x.reply
-                )
+                    str(x.chat_id), x.keyword, None, file_type.value, x.reply)
 
             SESSION.add(filt)
             SESSION.commit()
@@ -402,10 +388,8 @@ def __migrate_filters():
 def migrate_chat(old_chat_id, new_chat_id):
     with CUST_FILT_LOCK:
         chat_filters = (
-            SESSION.query(CustomFilters)
-            .filter(CustomFilters.chat_id == str(old_chat_id))
-            .all()
-        )
+            SESSION.query(CustomFilters).filter(
+                CustomFilters.chat_id == str(old_chat_id)).all())
         for filt in chat_filters:
             filt.chat_id = str(new_chat_id)
         SESSION.commit()
@@ -416,8 +400,8 @@ def migrate_chat(old_chat_id, new_chat_id):
 
         with BUTTON_LOCK:
             chat_buttons = (
-                SESSION.query(Buttons).filter(Buttons.chat_id == str(old_chat_id)).all()
-            )
+                SESSION.query(Buttons).filter(
+                    Buttons.chat_id == str(old_chat_id)).all())
             for btn in chat_buttons:
                 btn.chat_id = str(new_chat_id)
             SESSION.commit()
